Created SelectDateRange
SelectDateRange.dart uses flutter_date_pickers
package to display the calendar, for selecting
range of dates.

The onChanged event of the DateRangePicker has
been handled, to set the states startDate, and
endDate.

The states are then used in the UI, using
DateFormat (part of intl.dart package);

The states are also sent to the next page by
passing as parameters.