Updated HotelsAvailable.dart file
Improved the UI for the appbar
Designed the UI for displaying the hotels
Defined a list of hotels, in the state
of the HotelsAvailable page

Used it to populate the hotel display cards

There is also the facility to search on the
basis of rooms. In the previous screen,
number of rooms is selected, the state is passed
as parameter. Now, among the list of hotels
(list of map objects), we get only those hotels,
which have more rooms than the search query.

So, if the search query requires 4 rooms, then
there are no hotels displayed, as the list
has 3 or 0 rooms in each.